extract InstallerWorker
